feat(caws-environments): rebuild environment definition
